first release push

+++ angular-11-client/src/app/components/admin-view/admin-view.component.ts
@@ -5,9 +5,9 @@ import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
 
 @Component({
-  selector: 'app-users-list',
-  templateUrl: './users-list.component.html',
-  styleUrls: ['./users-list.component.css']
+  selector: 'app-admin-view',
+  templateUrl: './admin-view.component.html',
+  styleUrls: ['./admin-view.component.css']
 })
 export class UsersListComponent implements OnInit {
   users?: User[];
@@ -16,12 +16,14 @@ export class UsersListComponent implements OnInit {
   title = '';
   message = '';
   isAdmin: boolean = false;
+  activeUsers: number = 0;
+  votedUsers: number = 0;
+
 
   constructor(private UserService: UserService,  private router: Router) { }
 
   ngOnInit(): void {
     this.getGrants();
-
   }
 
   retrieveUsers(): void {
@@ -34,7 +36,8 @@ export class UsersListComponent implements OnInit {
       .subscribe(
         data => {
           this.users = data;
-          console.log(data);
+          console.log('[retrieveUsers]',data);
+          this.loadStatistics();
         },
         error => {
           console.log(error);
@@ -51,15 +54,19 @@ export class UsersListComponent implements OnInit {
       mail: localStorage.getItem('mail') || '',
        sessionId: localStorage.getItem('sessionId') || '',
     }
-    this.UserService.deleteAll(data)
-      .subscribe(
-        response => {
-          console.log(response);
-          this.retrieveUsers();
-        },
-        error => {
-          console.log(error);
-        });
+    if (confirm('Are you sure, you want to delete all users? The calling user will remain.')){
+      this.UserService.deleteAll(data)
+        .subscribe(
+          response => {
+            console.log(response);
+            this.retrieveUsers();
+          },
+          error => {
+            console.log(error);
+          });
+    }else {
+      this.retrieveUsers();
+    }
   }
 
   resetHasVoted(): void {
@@ -67,16 +74,43 @@ export class UsersListComponent implements OnInit {
       mail: localStorage.getItem('mail') || '',
        sessionId: localStorage.getItem('sessionId') || '',
     }
-    this.UserService.resetVoting(data, false)
-      .subscribe(
-        response => {
-          console.log(response);
-          this.retrieveUsers();
-        },
-        error => {
-          console.log(error);
-        });
+    if (confirm('Are you sure, you want to reset hasVoted for users? All users can vote again afterwards!')){
+      this.UserService.resetVoting(data, false)
+        .subscribe(
+          response => {
+            console.log(response);
+            this.retrieveUsers();
+          },
+          error => {
+            console.log(error);
+          });
+    } else {
+       this.retrieveUsers();
+    }
   }
+
+  updateList(): void {
+    const data = {
+      callingUser:{
+        mail: localStorage.getItem('mail') || '',
+        sessionId: localStorage.getItem('sessionId') || '',
+      },
+      users: this.users
+      }
+      if (confirm('Are you sure, you want to update all users? ')){
+      this.UserService.updateList(data)
+        .subscribe(
+          response => {
+            console.log(response);
+            this.retrieveUsers();
+          },
+          error => {
+            console.log(error);
+          });
+      } else {
+           this.retrieveUsers();
+      }
+    }
 
   getGrants(): void {
     const data = {
@@ -101,6 +135,21 @@ export class UsersListComponent implements OnInit {
             console.log(error);
             this.isAdmin = false;
           });
+    }
+
+    loadStatistics(): void {
+        this.activeUsers = 0;
+        this.votedUsers = 0;
+        if(this.users){
+          for(let i = 0; i < this.users.length; i++){
+            if(this.users && this.users[i].hasVoted){
+              this.votedUsers++;
+            }
+            if(this.users && this.users[i].active){
+              this.activeUsers++;
+            }
+          }
+        }
     }
 
 }
